BACKEND-75: separate CompetitionController and CompetitionTeamController; fix endponts
