+++ backend/app/web/src/main/kotlin/ru/nekoguys/game/web/controller/CompetitionController.kt
@@ -1,18 +1,13 @@
 package ru.nekoguys.game.web.controller
 
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
-import org.springframework.http.codec.ServerSentEvent
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.*
 import ru.nekoguys.game.web.dto.*
 import ru.nekoguys.game.web.service.CompetitionService
-import ru.nekoguys.game.web.util.toOkResponse
 import ru.nekoguys.game.web.util.toResponseEntity
 import ru.nekoguys.game.web.util.withMDCContext
-import ru.nekoguys.game.web.util.withRequestIdInContext
 import java.security.Principal
 import javax.validation.Valid
 
@@ -53,60 +48,8 @@ class CompetitionController(
                 userEmail = principal.name,
                 limit = amount,
                 offset = start,
-            ).toOkResponse()
+            ).let { ResponseEntity.ok(it) }
         }
-
-    @PostMapping(
-        "/create_team",
-        consumes = [MediaType.APPLICATION_JSON_VALUE],
-        produces = [MediaType.APPLICATION_JSON_VALUE],
-    )
-    @PreAuthorize("hasRole('STUDENT')")
-    suspend fun createTeam(
-        principal: Principal,
-        @RequestBody request: CreateTeamRequest,
-    ): ResponseEntity<out CreateTeamResponse> =
-        withMDCContext {
-            competitionService.createTeam(
-                studentEmail = principal.name,
-                request = request,
-            ).toResponseEntity()
-        }
-
-    @PostMapping(
-        "/join_team",
-        consumes = [MediaType.APPLICATION_JSON_VALUE],
-        produces = [MediaType.APPLICATION_JSON_VALUE],
-    )
-    @PreAuthorize("hasRole('STUDENT')")
-    suspend fun joinTeam(
-        principal: Principal,
-        @RequestBody request: JoinTeamRequest,
-    ): ResponseEntity<out JoinTeamResponse> =
-        withMDCContext {
-            competitionService.joinTeam(
-                studentEmail = principal.name,
-                request = request,
-            ).toResponseEntity()
-        }
-
-    @RequestMapping(
-        "/team_join_events/{pin}",
-        produces = [MediaType.TEXT_EVENT_STREAM_VALUE],
-    )
-    fun teamJoinEvents(
-        principal: Principal,
-        @PathVariable pin: String,
-    ): Flow<ServerSentEvent<TeamUpdateNotification>> =
-        competitionService
-            .teamJoinMessageFlow(
-                userEmail = principal.name,
-                sessionPin = pin,
-            )
-            .map { message ->
-                ServerSentEvent.builder(message).id("teamStream").build()
-            }
-            .withRequestIdInContext()
 
     @PostMapping(
         "/check_pin",
@@ -123,14 +66,16 @@ class CompetitionController(
             .toResponseEntity()
 
     @GetMapping(
-        "/get_clone_info/{pin}",
+        "/get_clone_info/{sessionPin}",
         produces = [MediaType.APPLICATION_JSON_VALUE]
     )
     @PreAuthorize("hasRole('TEACHER')")
-    suspend fun getCompetitionInfo(@PathVariable pin: String) {
+    suspend fun getCompetitionInfo(
+        @PathVariable sessionPin: String,
+    ) {
         withMDCContext {
             competitionService
-                .getCompetitionCloneInfo(pin)
+                .getCompetitionCloneInfo(sessionPin)
                 .toResponseEntity()
         }
     }
