+++ backend/app/web/src/test/kotlin/ru/nekoguys/game/web/controller/process/StartCompetitionControllerTest.kt
@@ -31,12 +31,12 @@ class StartCompetitionControllerTest @Autowired constructor(
     @Test
     @WithMockUser(username = TestGame.DEFAULT_EMAIL, roles = ["TEACHER"])
     fun `can start competition`() {
-        val competitionPin = game.createCompetition()
-        repeat(2) { game.createTeam(competitionPin) }
+        val sessionPin = game.createSession()
+        repeat(2) { game.createTeam(sessionPin) }
 
         webTestClient
             .get()
-            .uri("/api/competition_process/$competitionPin/start_competition")
+            .uri("/api/competition_process/$sessionPin/start_competition")
             .exchange()
             .expectStatus().isOk
             .expectBody()
@@ -46,12 +46,12 @@ class StartCompetitionControllerTest @Autowired constructor(
     @Test
     @WithMockUser(username = "test-student@hse.ru", roles = ["STUDENT"])
     fun `student can't start competition`() {
-        val competitionPin = game.createCompetition()
-        repeat(2) { game.createTeam(competitionPin) }
+        val sessionPin = game.createSession()
+        repeat(2) { game.createTeam(sessionPin) }
 
         webTestClient
             .get()
-            .uri("/api/competition_process/$competitionPin/start_competition")
+            .uri("/api/competition_process/$sessionPin/start_competition")
             .exchange()
             .expectStatus().isForbidden
     }
@@ -59,13 +59,13 @@ class StartCompetitionControllerTest @Autowired constructor(
     @Test
     @WithMockUser(username = TestGame.DEFAULT_EMAIL, roles = ["TEACHER"])
     fun `can't start draft competition`() {
-        val competitionPin = game.createCompetition(
+        val sessionPin = game.createSession(
             request = TestGame.DEFAULT_CREATE_DRAFT_COMPETITION_REQUEST
         )
 
         webTestClient
             .get()
-            .uri("/api/competition_process/$competitionPin/start_competition")
+            .uri("/api/competition_process/$sessionPin/start_competition")
             .exchange()
             .expectStatus().isBadRequest
             .expectBody()
@@ -75,13 +75,13 @@ class StartCompetitionControllerTest @Autowired constructor(
     @Test
     @WithMockUser(username = TestGame.DEFAULT_EMAIL, roles = ["TEACHER"])
     fun `can start competition twice`() {
-        val competitionPin = game.createCompetition()
-        repeat(2) { game.createTeam(competitionPin) }
-        game.startCompetition(competitionPin)
+        val sessionPin = game.createSession()
+        repeat(2) { game.createTeam(sessionPin) }
+        game.startCompetition(sessionPin)
 
         webTestClient
             .get()
-            .uri("/api/competition_process/$competitionPin/start_competition")
+            .uri("/api/competition_process/$sessionPin/start_competition")
             .exchange()
             .expectStatus().isBadRequest
             .expectBody()
