+++ backend/app/web/src/test/kotlin/ru/nekoguys/game/web/util/TestGame.kt
@@ -15,11 +15,13 @@ import ru.nekoguys.game.entity.user.repository.load
 import ru.nekoguys.game.web.dto.*
 import ru.nekoguys.game.web.service.CompetitionProcessService
 import ru.nekoguys.game.web.service.CompetitionService
+import ru.nekoguys.game.web.service.CompetitionTeamService
 import java.util.concurrent.atomic.AtomicLong
 
 @Component
 class TestGame(
     private val competitionService: CompetitionService,
+    private val competitionTeamService: CompetitionTeamService,
     private val competitionProcessService: CompetitionProcessService,
     private val competitionSessionRepository: CompetitionSessionRepository,
     private val pinGenerator: SessionPinDecoder,
@@ -30,7 +32,7 @@ class TestGame(
     private fun nextEmail(role: UserRole) =
         "${role.toString().lowercase()}-${nextIndex()}@hse.ru"
 
-    private fun nextCompetitionName() =
+    private fun nextSessionName() =
         "Competition ${nextIndex()}"
 
     private fun nextTeamName() = "Team ${nextIndex()}"
@@ -43,18 +45,18 @@ class TestGame(
         userRepository.create(email, "{noop}$password", role)
     }
 
-    fun loadCompetitionSession(
-        pin: String = createCompetition(),
+    fun loadSession(
+        pin: String = createSession(),
     ): CompetitionSession.Full = runBlocking {
         val sessionId = pinGenerator.decodeIdFromPin(pin)
         competitionSessionRepository
             .load(sessionId!!, CompetitionSession.Full)
     }
 
-    fun createAndLoadCompetition(
+    fun createAndLoadSession(
         teacher: User = createUser(UserRole.Teacher),
         request: CreateCompetitionRequest = DEFAULT_CREATE_COMPETITION_REQUEST
-            .copy(name = nextCompetitionName()),
+            .copy(name = nextSessionName()),
         additionalActions: suspend (String) -> Unit = {},
     ): CompetitionSession.Full = runBlocking {
         competitionService.create(teacher.email, request)
@@ -73,65 +75,63 @@ class TestGame(
         competitionSessionRepository.load(sessionId, CompetitionSession.Full)
     }
 
-    fun createCompetition(
+    fun createSession(
         teacher: User = createUser(UserRole.Teacher),
         request: CreateCompetitionRequest = DEFAULT_CREATE_COMPETITION_REQUEST
-            .copy(name = nextCompetitionName()),
+            .copy(name = nextSessionName()),
     ): String =
-        createAndLoadCompetition(teacher, request)
+        createAndLoadSession(teacher, request)
             .id
             .toPin()
 
     fun createTeam(
-        competitionPin: String = createCompetition(),
+        sessionPin: String = createSession(),
         captain: User = createUser(UserRole.Student),
         teamName: String = nextTeamName(),
         password: String = DEFAULT_PASSWORD,
     ): TestCreateTeamResult {
         val request = CreateTeamRequest(
-            gameId = competitionPin,
             teamName = teamName,
             captainEmail = captain.email,
             password = password,
         )
 
         runBlocking {
-            val response = competitionService.createTeam(captain.email, request)
+            val response = competitionTeamService.create(sessionPin, captain.email, request)
             check(response is CreateTeamResponse.Success)
         }
 
         return TestCreateTeamResult(
-            competitionPin = request.gameId,
+            sessionPin = sessionPin,
             teamName = request.teamName,
             password = request.password,
         )
     }
 
     data class TestCreateTeamResult(
-        val competitionPin: String,
+        val sessionPin: String,
         val teamName: String,
         val password: String,
     )
 
     fun joinTeam(
-        competitionPin: String = createCompetition(),
-        teamName: String = createTeam(competitionPin = competitionPin).teamName,
+        sessionPin: String = createSession(),
+        teamName: String = createTeam(sessionPin = sessionPin).teamName,
         teamMember: User = createUser(role = UserRole.Student),
         password: String = DEFAULT_PASSWORD,
     ): TestCreateTeamResult {
         val request = JoinTeamRequest(
-            competitionPin = competitionPin,
             teamName = teamName,
             password = password,
         )
 
         runBlocking {
-            val response = competitionService.joinTeam(teamMember.email, request)
+            val response = competitionTeamService.join(sessionPin, teamMember.email, request)
             check(response is JoinTeamResponse.Success)
         }
 
         return TestCreateTeamResult(
-            competitionPin = request.competitionPin,
+            sessionPin = sessionPin,
             teamName = request.teamName,
             password = request.password,
         )
@@ -143,19 +143,19 @@ class TestGame(
      * С ним тесты иногда проще читать
      */
     fun createAndJoinTeam(
-        competitionPin: String = createCompetition(),
+        sessionPin: String = createSession(),
         teamMember: User = createUser(role = UserRole.Student),
     ): TestCreateTeamResult = runBlocking {
-        joinTeam(competitionPin, teamMember = teamMember)
+        joinTeam(sessionPin, teamMember = teamMember)
     }
 
     fun startCompetition(
-        competitionPin: String = createCompetition(),
+        sessionPin: String = createSession(),
     ): Unit = runBlocking {
-        val teacherId = loadCompetitionSession(competitionPin).creatorId
+        val teacherId = loadSession(sessionPin).creatorId
         val teacher = userRepository.load(teacherId)
         val response = competitionProcessService
-            .startCompetition(teacher.email, competitionPin)
+            .startCompetition(teacher.email, sessionPin)
         check(response == StartCompetitionResponse.Success)
     }
